Auto encoder util functions, and other some minor problems solved. Also, cross entropy error function assert insertion to control output and activation values to be [0,1]

+++ packages/ann/ann_base/c_src/bias_connection.cc
@@ -83,7 +83,8 @@ namespace ANN {
 					    unsigned int first_weight_pos,
 					    unsigned int column_size) {
     if ((total_size + (column_size - num_inputs)*num_outputs +
-	 first_weight_pos) > data->size) ERROR_EXIT(24, "Incorrect matrix size\n");
+	 first_weight_pos) > static_cast<unsigned int>(data->size))
+      ERROR_EXIT(24, "Incorrect matrix size\n");
     
     unsigned int current_w_pos = first_weight_pos;
     float *w                   = weights->getPPALForReadAndWrite();
@@ -101,7 +102,8 @@ namespace ANN {
 					      unsigned int first_weight_pos,
 					      unsigned int column_size) {
     if ((total_size + (column_size - num_inputs)*num_outputs +
-	 first_weight_pos) > data->size) ERROR_EXIT(24, "Incorrect matrix size\n");
+	 first_weight_pos) > static_cast<unsigned int>(data->size))
+      ERROR_EXIT(24, "Incorrect matrix size\n");
     
     unsigned int current_w_pos = first_weight_pos;
     const float *w             = weights->getPPALForRead();
